requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
black==24.8.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.1
decorator==5.1.1
Django==5.0.7
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
executing==2.0.1
idna==3.7
inflection==0.5.1
ipython==8.26.0
isort==5.13.2
jedi==0.19.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
stripe==10.8.0
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
